requirements.txt
django
pyTelegramBotAPI
bs4
lxml
django-jet
django-ckeditor